hackernews-howtographql/requirements.txt
aniso8601==7.0.0
autopep8==1.4.4
Django==2.2.5
django-environ==0.4.5
django-filter==2.2.0
django-graphql-jwt==0.2.2
graphene==2.1.8
graphene-django==2.5.0
graphql-core==2.2.1
graphql-relay==2.0.0
promise==2.2.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.2
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
